requirements.txt
selenium
beautifulsoup4
google-genai
yaspin